Refactor/#210 Tastinginfo, LocationInfo 타입 위치 변경 (#211)
* refactor: LocationInfo 위치 constants로 이동
* refactor: TastingInfo 위치 constants로 이동
