+++ src/entities/meetings/model/types.ts
@@ -1,4 +1,4 @@
-import { DrinkType } from '@/shared';
+import { DrinkType, LocationInfo } from '@/shared';
 
 export enum SortType {
   latest = 'LATEST',
@@ -20,11 +20,7 @@ export interface MeetingCardInfo {
   meetingId: number;
   meetingAuthor: string;
   meetingTitle: string;
-  location: {
-    sido: string;
-    address: string;
-    detail: string;
-  };
+  location: LocationInfo;
   participationFee: number;
   startAt: string;
   joinEndAt: string;
