+++ src/shared/constants/types.ts
@@ -4,3 +4,19 @@ export enum DrinkType {
   wine = 'WINE',
   end = 'END',
 }
+
+export interface LocationInfo {
+  sido: string;
+  address: string;
+  detail: string;
+}
+
+export interface TastingInfo {
+  drinkId?: number;
+  drinkName?: string;
+  drinkTaste?: string; // 음료 맛
+  drinkFlavor?: string; // 음료 향
+  drinkColor?: string; // 음료 색
+  drinkImgUrl?: string;
+  drinkType?: DrinkType;
+}
