+++ src/shared/ui/form/AddressField.tsx
@@ -4,7 +4,8 @@ import { useState } from 'react';
 import DaumPostcode, { Address } from 'react-daum-postcode';
 import useFieldValue from '../../lib/form/model/hooks/useFieldValue';
 import FormField from './FormField';
-import { LocationInfo, AddressFieldProps } from '../../lib/form/model/types';
+import { LocationInfo } from '../../constants';
+import { AddressFieldProps } from '../../lib/form/model/types';
 import { Button, Input, Popover, PopoverContent, PopoverTrigger } from '../index';
 
 export default function AddressField({
