+++ src/widgets/post-meeting/model/types.ts
@@ -1,13 +1,8 @@
-import { DrinkType } from '@/shared';
-import { TastingInfo } from '@/entities/tasting-list';
+import { DrinkType, LocationInfo, TastingInfo } from '@/shared';
 
 export default interface PostMeetingFormData {
   meetingTitle: string;
-  location: {
-    sido: string;
-    address: string;
-    detail: string;
-  };
+  location: LocationInfo;
   participationFee: number;
   startAt: string;
   joinEndAt: string;
